feat(#26): replace CostWidget with StatsWidget showing hp, mass, speed, raid cost
